Clarify internal dry-run arg name
I want the “enable destructive mode” argument to have a positive name
(`—enact`) but it was probably causing some confusion as to the default
state of the variable.

Flipping some of the bool logic around with the new name should make
this a bit clearer.

See #1, thanks to willmot for making me think about this.